raslen

+++ legacy/app/dashboard/products.tsx
@@ -2,16 +2,18 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import { Typography, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, IconButton, Button, TextField, Modal, Box } from '@mui/material';
 import { Delete, Edit } from '@mui/icons-material';
-interface Product{
-    id:number,
-    name:string,
-    price:number,
-    description:string,
-    category:string,
-    stock:number,
-    picture:string,
-    userId:number
- }
+import Swal from 'sweetalert2'
+
+interface Product {
+  id: number;
+  name: string;
+  price: number;
+  description: string;
+  category: string;
+  stock: number;
+  picture: string;
+  userId: number;
+}
 
 const styles = {
   header: {
@@ -33,17 +35,17 @@ const styles = {
   },
   tableBodyRow: {
     '&:nth-of-type(odd)': {
-      backgroundColor: '#e8eaf6',
+      backgroundColor: '#f3e5f5',
     },
   },
   modal: {
-    position: 'absolute',
+    position: 'absolute' as 'absolute',
     top: '50%',
     left: '50%',
     transform: 'translate(-50%, -50%)',
     width: 500,
     bgcolor: '#ffffff',
-    boxShadow: 2434,
+    boxShadow: 24,
     p: 4,
     borderRadius: '10px',
   },
@@ -52,20 +54,36 @@ const styles = {
   },
   button: {
     marginTop: '15px',
-    backgroundColor: '#c42121',
+    backgroundColor: '#5b3594',
     color: '#ffffff',
     '&:hover': {
-      backgroundColor: '#0d0d0f',
+      backgroundColor: '#452780',
+    },
+  },
+  actionButton: {
+    marginLeft: '10px',
+    backgroundColor: '#5e35b1',
+    color: '#ffffff',
+    '&:hover': {
+      backgroundColor: '#452780',
+    },
+  },
+  addButton: {
+    marginBottom: '20px',
+    backgroundColor: '#5b3594',
+    color: '#ffffff',
+    '&:hover': {
+      backgroundColor: '#452780',
     },
   },
 };
 
-const ProductsTable:React.FC = () => {
+const ProductsTable: React.FC = () => {
   const [products, setProducts] = useState<Product[]>([]);
   const [modify, setModify] = useState<boolean>(false);
   const [addModalOpen, setAddModalOpen] = useState(false);
   const [currentProduct, setCurrentProduct] = useState({
-    id:0,
+    id: 0,
     name: "",
     price: 0,
     picture: "",
@@ -80,34 +98,51 @@ const ProductsTable:React.FC = () => {
     category: "",
     stock: 0,
     description: "",
-    userId:0
+    userId: 0,
   });
-  
+
   useEffect(() => {
     fetchProducts();
   }, []);
 
   const fetchProducts = async () => {
     try {
-      const response = await axios.get< Product[]>("http://localhost:5000/Admin/products");
+      const response = await axios.get<Product[]>("http://localhost:5000/Admin/products");
       setProducts(response.data);
     } catch (error) {
       console.error("Error fetching products:", error);
     }
   };
 
-  const deleteProduct = async (productId:number) => {
-    try {
-      await axios.delete< Product[]>(`http://localhost:5000/admin/products/${productId}`);
-      setProducts(products.filter((product) => product.id !== productId));
-    } catch (error) {
-      console.error("Error deleting product:", error);
-    }
+  const deleteProduct = async (productId: number) => {
+  
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this product!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d32f2f',
+      cancelButtonColor: '#5e35b1',
+      confirmButtonText: 'Yes, delete it!',
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          await axios.delete<Product[]>(`http://localhost:5000/admin/products/${productId}`);
+          setProducts(products.filter((product) => product.id !== productId));
+       
+          Swal.fire('Deleted!', 'Your product has been deleted.', 'success');
+        } catch (error) {
+          console.error("Error deleting product:", error);
+       
+          Swal.fire('Error!', 'Failed to delete the product.', 'error');
+        }
+      }
+    });
   };
 
   const updateProduct = async () => {
     try {
-      await axios.put< Product[]>(
+      await axios.put<Product[]>(
         `http://localhost:5000/admin/products/${currentProduct.id}`,
         currentProduct
       );
@@ -116,21 +151,20 @@ const ProductsTable:React.FC = () => {
     } catch (error) {
       console.error("Error updating product:", error);
     }
-  }
-  
+  };
+
   const addProduct = async () => {
-      try {
+    try {
       const { userId, ...productData } = newProduct;
-      await axios.post< Product[]>("http://localhost:5000/admin/products/add", productData);
-            fetchProducts()
-      handleCloseAdd()
+      await axios.post<Product[]>("http://localhost:5000/admin/products/add", productData);
+      fetchProducts();
+      handleCloseAdd();
     } catch (error) {
       console.error("Error adding product:", error);
     }
-  }
+  };
 
-
-  const handleOpenModify = (product:Product) => {
+  const handleOpenModify = (product: Product) => {
     setCurrentProduct(product);
     setModify(true);
   };
@@ -148,13 +182,12 @@ const ProductsTable:React.FC = () => {
     });
   };
 
-
   const handleOpenAdd = () => {
-    setAddModalOpen(true)
+    setAddModalOpen(true);
   };
 
   const handleCloseAdd = () => {
-    setAddModalOpen(false)
+    setAddModalOpen(false);
     setNewProduct({
       name: "",
       price: 0,
@@ -162,7 +195,7 @@ const ProductsTable:React.FC = () => {
       category: "",
       stock: 0,
       description: "",
-      userId:0
+      userId: 0,
     });
   };
 
@@ -171,7 +204,7 @@ const ProductsTable:React.FC = () => {
       <Typography variant="h3" gutterBottom sx={styles.header}>
         Products
       </Typography>
-      <Button variant="contained" sx={styles.button} onClick={handleOpenAdd}>
+      <Button variant="contained" sx={styles.addButton} onClick={handleOpenAdd}>
         Add Product
       </Button>
       <TableContainer>
@@ -205,7 +238,6 @@ const ProductsTable:React.FC = () => {
                   <IconButton onClick={() => deleteProduct(product.id)}>
                     <Delete sx={{ color: '#d32f2f' }} />
                   </IconButton>
-                  
                 </TableCell>
               </TableRow>
             ))}
@@ -229,8 +261,9 @@ const ProductsTable:React.FC = () => {
             fullWidth
             label="Price"
             margin="normal"
+            type="number"
             value={currentProduct.price}
-            onChange={(e) => setCurrentProduct({ ...currentProduct, price: parseInt(e.target.value) })}
+            onChange={(e) => setCurrentProduct({ ...currentProduct, price: parseFloat(e.target.value) })}
             sx={styles.input}
           />
           <TextField
@@ -253,8 +286,9 @@ const ProductsTable:React.FC = () => {
             fullWidth
             label="Stock"
             margin="normal"
+            type="number"
             value={currentProduct.stock}
-            onChange={(e) => setCurrentProduct({ ...currentProduct, stock:  parseInt(e.target.value)})}
+            onChange={(e) => setCurrentProduct({ ...currentProduct, stock: parseInt(e.target.value) })}
             sx={styles.input}
           />
           <TextField
@@ -267,7 +301,7 @@ const ProductsTable:React.FC = () => {
           />
           <Button variant="contained" sx={styles.button} onClick={updateProduct}>
             Update Product
-          </Button> 
+          </Button>
         </Box>
       </Modal>
       <Modal open={addModalOpen} onClose={handleCloseAdd}>
@@ -287,8 +321,9 @@ const ProductsTable:React.FC = () => {
             fullWidth
             label="Price"
             margin="normal"
+            type="number"
             value={newProduct.price}
-            onChange={(e) => setNewProduct({ ...newProduct, price:  parseInt(e.target.value) })}
+            onChange={(e) => setNewProduct({ ...newProduct, price: parseFloat(e.target.value) })}
             sx={styles.input}
           />
           <TextField
@@ -311,8 +346,9 @@ const ProductsTable:React.FC = () => {
             fullWidth
             label="Stock"
             margin="normal"
+            type="number"
             value={newProduct.stock}
-            onChange={(e) => setNewProduct({ ...newProduct, stock:  parseInt(e.target.value) })}
+            onChange={(e) => setNewProduct({ ...newProduct, stock: parseInt(e.target.value) })}
             sx={styles.input}
           />
           <TextField
@@ -325,11 +361,11 @@ const ProductsTable:React.FC = () => {
           />
           <Button variant="contained" sx={styles.button} onClick={addProduct}>
             Add Product
-          </Button> 
+          </Button>
         </Box>
       </Modal>
     </div>
-  )
-}
+  );
+};
 
 export default ProductsTable;
